UnPWAified the stats page
